fixed issue of extra white-space in the first row
the grid row was using min-content. the first column spanned two rows and so I had assumed the min-content would be that of only the second column which occupied only the first row.

the problem was the min-content was minimally bounded by the 6px row-gap in the first column's flex container