simplify state setup
prior to this there was a separate state for the 'working markdown'. i did this in order to prevent copying the entire markdown array every time the textarea is edited. this had downsides such as more convoluted code and necessitated keeping in sync with the markdown array.

i will try to optimise in a future update. it is currently fine but will not scale as the array becomes larger