update handling of menu file items.
state has been moved down into the menu component (as this is the only place they are used).

state was previously higher in the component tree as i was using a more naive state implementation of storing the whole database in one state variable. as state is now more broken up, I can move its consituent parts closer to where they are used.

an event handler is used in a custom hook to update the file list whenever the database (in local storage) is updated.