fix adding of new files with new state structure
previously, the built in 'storage' event was being used to trigger updates to the file list state. it did not work as i intended and so now a custom event 'storageUpdated' is used.

my misunderstanding was with the 'storage' event. it is only triggered when another window/tab modifies the storage - not the current one.