moved 'markdown' state back up into 'App' component
the 'Header' component also needs access to the markdown state in order to be able to update version of file in local storage with the current, edited content.

had to add a useEffect in 'App' component to refresh the working markdown state when current index state is changed. i was trying to avoid this as it will cause an extra render.