fixed theming for dialog
there was an issue with twMerge confusing my custom classes. it was assuming text-orange and text-preview-h4 to be part of the same 'class group' - i.e. it thought they were both altering the color property and so one was overriding the other unintentionally.

i have configured twMerge using extendTailwindMerge as per the docs.